Use common sentry interface
