+++ src/common/store/errors/actions.ts
@@ -1,16 +1,7 @@
+import { Level } from './level'
+import { AdditionalInfo, ReportToSentryArgs } from './reportToSentry'
+
 export const HANDLE_ERROR = 'ERROR/HANDLE_ERROR'
-
-export enum Level {
-  'Critical' = 'Critical',
-  'Warning' = 'Warning',
-  'Fatal' = 'Fatal',
-  'Debug' = 'Debug',
-  'Error' = 'Error',
-  'Info' = 'Info',
-  'Log' = 'Log'
-}
-
-export type AdditionalInfo = Record<string, unknown>
 
 export type HandleErrorAction = {
   type: typeof HANDLE_ERROR
@@ -27,15 +18,11 @@ export type HandleErrorAction = {
 type HandleActions = HandleErrorAction
 
 type HandleErrorArgs = {
-  name?: string
-  message: string
   shouldRedirect: boolean
   shouldReport?: boolean
   shouldToast?: boolean
-  additionalInfo?: AdditionalInfo
-
-  level?: Level
-}
+  message: string
+} & Omit<ReportToSentryArgs, 'error'>
 
 export const handleError = ({
   name,
