+++ src/common/store/errors/level.ts
@@ -1,0 +1,9 @@
+export enum Level {
+  'Critical' = 'Critical',
+  'Warning' = 'Warning',
+  'Fatal' = 'Fatal',
+  'Debug' = 'Debug',
+  'Error' = 'Error',
+  'Info' = 'Info',
+  'Log' = 'Log'
+}
