+++ src/services/AudioManager.ts
@@ -3,6 +3,7 @@ import { put, call, fork, select } from 'redux-saga/effects'
 import { Name } from 'common/models/Analytics'
 import { User } from 'common/models/User'
 import * as errorActions from 'common/store/errors/actions'
+import { Level } from 'common/store/errors/level'
 import { getModalVisibility, setVisibility } from 'common/store/ui/modals/slice'
 import { Nullable } from 'common/utils/typeUtils'
 import AudiusBackend from 'services/AudiusBackend'
@@ -106,7 +107,7 @@ class AudioManager {
           shouldRedirect: false,
           shouldReport: true,
           additionalInfo: { errorMessage: error.message },
-          level: errorActions.Level.Critical
+          level: Level.Critical
         })
       )
     }
